+++ crates/gbnf/assets/enhance-auto.gbnf
