+++ crates/llama/Cargo.toml
@@ -3,23 +3,9 @@ name = "llama"
 version = "0.1.0"
 edition = "2021"
 
-[dev-dependencies]
-hypr-buffer = { workspace = true }
-hypr-data = { workspace = true }
-hypr-listener-interface = { workspace = true }
-hypr-template = { workspace = true }
-hypr-timeline = { workspace = true }
-
-colored = "3.0.0"
-dirs = { workspace = true }
-indoc = "2.0.6"
-minijinja = { workspace = true }
-rand = "0.9.0"
-serde_json = { workspace = true }
-
 [dependencies]
+hypr-gbnf = { workspace = true }
 hypr-gguf = { workspace = true }
-include_url_macro = { workspace = true }
 
 encoding_rs = "0.8.35"
 gbnf-validator = { workspace = true }
@@ -37,3 +23,14 @@ llama-cpp-2 = { git = "https://github.com/utilityai/llama-cpp-rs", default-featu
 
 [target.'cfg(target_os = "macos")'.dependencies]
 llama-cpp-2 = { git = "https://github.com/utilityai/llama-cpp-rs", features = ["openmp", "native", "metal"], branch = "update-llama-cpp-2025-04-06" }
+
+[dev-dependencies]
+hypr-buffer = { workspace = true }
+hypr-data = { workspace = true }
+hypr-listener-interface = { workspace = true }
+hypr-template = { workspace = true }
+hypr-timeline = { workspace = true }
+
+dirs = { workspace = true }
+rand = "0.9.0"
+serde_json = { workspace = true }
